# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Order Management System with Hexagonal Architecture.",
        "title": "OMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/orders": {
            "post": {
                "description": "Creates an order with a random customer ID and specific price",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order Details",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.Order"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Returns a single order by its uint64 ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Get an order",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Order"
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Order": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.OrderItem"
                    }
                },
                "status": {
                    "$ref": "#/definitions/entities.OrderStatus"
                },
                "totalPrice": {
                    "type": "integer",
                    "format": "int64"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "entities.OrderItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "price": {
                    "type": "integer",
                    "format": "int64"
                },
                "quantity": {
                    "type": "integer",
                    "format": "int32"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entities.OrderStatus": {
            "type": "integer",
            "format": "int32",
            "enum": [
                0,
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "Pending",
                "Paid",
                "Shipped",
                "Cancelled"
            ]
        },
        "order.CreateOrderRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "price"
            ],
            "properties": {
                "customer_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                }
            }
        }
    }
}