# internal/handlers/http/order/handler.go
package order

import (
	"encoding/json"
	"net/http"
	"strconv"
	"strings"

	"github.com/hassan-alidoost/oms-project/internal/domain/entities"
	"github.com/hassan-alidoost/oms-project/internal/services"
)

type OrderHandler struct {
	service *services.OrderService
}

func NewOrderHandler(service *services.OrderService) *OrderHandler {
	return &OrderHandler{service: service}
}

func (h *OrderHandler) RegisterRoutes(mux *http.ServeMux) {
	mux.HandleFunc("POST /orders", h.CreateOrder)
	mux.HandleFunc("GET /orders/{id}", h.GetOrder)
}

// CreateOrder handles the creation of a new order
// @Summary Create a new order
// @Description Creates an order with a random customer ID and specific price
// @Tags orders
// @Accept json
// @Produce json
// @Param order body CreateOrderRequest true "Order Details"
// @Success 201 {object} entities.Order
// @Failure 400 {string} string "Invalid request"
// @Router /orders [post]
func (h *OrderHandler) CreateOrder(w http.ResponseWriter, r *http.Request) {
	var req CreateOrderRequest

	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	order, err := h.service.PlaceOrder(r.Context(), entities.EntityId(req.UserID), entities.Price(req.Price))
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(order)
}

// GetOrder retrieves an order by ID
// @Summary Get an order
// @Description Returns a single order by its uint64 ID
// @Tags orders
// @Produce json
// @Param id path uint64 true "Order ID"
// @Success 200 {object} entities.Order
// @Failure 404 {string} string "Order not found"
// @Router /orders/{id} [get]
func (h *OrderHandler) GetOrder(w http.ResponseWriter, r *http.Request) {
	parts := strings.Split(r.URL.Path, "/")
	if len(parts) < 3 {
		http.Error(w, "ID required", http.StatusBadRequest)
		return
	}

	id, err := strconv.ParseUint(parts[2], 10, 64)
	if err != nil {
		http.Error(w, "Invalid ID", http.StatusBadRequest)
		return
	}

	order, err := h.service.FindByID(r.Context(), entities.EntityId(id))
	if err != nil {
		http.Error(w, "Order not found", http.StatusNotFound)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(order)
}